+++ SLPWallet/SLPTokenUTXO.swift
@@ -10,7 +10,7 @@ import Foundation
 
 public class SLPTokenUTXO: SLPUTXO {
     public var rawTokenQty: Int
-    public var tokenQty: Float?
+    public var tokenQty: Double?
     
     public init(_ txid: String, satoshis: Int, cashAddress: String, scriptPubKey: String, index: Int, rawTokenQty: Int) {
         self.rawTokenQty = rawTokenQty
